Change db_id type from int to unsigned long in meta_get_keys()
meta_open() takes an unsigned long for db_id, so for consistence's sake
meta_get_keys() should as well.

+++ nasfs_driver/meta_db.h
@@ -5,5 +5,5 @@ int meta_open(const char *, unsigned long, bool, int);
 void meta_close();
 int meta_get(char *, char **);
 int meta_put(char *, char *);
-int meta_get_keys(char ***, char ***, int, const char*);
+int meta_get_keys(char ***, char ***, unsigned long, const char*);
 void meta_free(char *, char **);
